+++ src/main/java/woowacourse/shoppingcart/ui/ControllerAdvice.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import woowacourse.auth.exception.AuthException;
-import woowacourse.shoppingcart.dto.ErrorResponse;
+import woowacourse.shoppingcart.dto.response.ErrorResponse;
 import woowacourse.shoppingcart.exception.ShoppingCartException;
 
 @RestControllerAdvice
