+++ src/test/java/woowacourse/shoppingcart/application/ProductServiceTest.java
@@ -23,7 +23,7 @@ import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.domain.customer.Customer;
 import woowacourse.shoppingcart.domain.customer.UserName;
-import woowacourse.shoppingcart.dto.ProductResponse;
+import woowacourse.shoppingcart.dto.response.ProductResponse;
 
 @ExtendWith(MockitoExtension.class)
 class ProductServiceTest {
